Mexican spanish support
Feel free to change it to "Latin American Spanish" or just "Spanish" if you prefer, but keep in mind there *may* be a few differences with other varieties, even with other latin american varieties. Also note that "y" (and) changes to "e" if the next word starts with i and that "o" (or) changes to "u" if the next word starts with o, but I don't know how to do that here.